# Improve test runner performance by restoring from CRaC checkpoint