--- a/Dockerfile
+++ b/Dockerfile
@@ -1,13 +1,14 @@
-FROM gradle:8.12-jdk21 AS build
+FROM bellsoft/liberica-runtime-container:jdk-21-crac-musl AS build
 
 WORKDIR /app
-COPY --chown=gradle:gradle . /app
-RUN gradle -i --stacktrace clean build
+COPY . /app
+RUN /app/gradlew -i --stacktrace clean build
 
-FROM eclipse-temurin:21
+FROM bellsoft/liberica-runtime-container:jdk-21-crac-musl
 
 WORKDIR /opt/test-runner
-COPY bin/run.sh bin/run.sh
+COPY bin/run-to-create-crac-checkpoint.sh bin/run-to-create-crac-checkpoint.sh
+COPY bin/run-restore-from-checkpoint.sh bin/run-restore-from-checkpoint.sh
 COPY --from=build /app/build/libs/java-test-runner.jar .
 
-ENTRYPOINT ["sh", "/opt/test-runner/bin/run.sh"]
+ENTRYPOINT ["sh", "/opt/test-runner/bin/run-to-create-crac-checkpoint.sh"]