--- a/src/main/java/com/exercism/Restore.java
+++ b/src/main/java/com/exercism/Restore.java
@@ -1,0 +1,11 @@
+package com.exercism;
+
+import java.io.IOException;
+
+public class Restore {
+
+	public static void main(String[] args) throws IOException {
+		new TestRunner(args[0], args[1], args[2]).run();
+	}
+
+}