@SleeplessByte 2025-03-13
> I assume that my changes currently break the [deploy](https://github.com/exercism/java-test-runner/blob/67629f072af02ffe40b27aa6f3fa41b71d188cb7/.github/workflows/deploy.yml)? Currently I'm using `exercism/java-test-runner-crac-restore` as image name and I introduced an [extra script](https://github.com/exercism/java-test-runner/pull/147/files#diff-dd84856e1e52c87a95138052460a01f091aabdb12abe52c8b96c265f0b5b1be5) for the container image build.

Why do you think your changes will break the deploy?