@fapdash 2025-03-13
@SleeplessByte The GH action builds the image directly from the Dockerfile, but with this change it has to call [bin/build-crac-checkpoint-image.sh](https://github.com/exercism/java-test-runner/blob/37de8233d2b4b7ae8c78200f67e54c0aaf874fac/bin/build-crac-checkpoint-image.sh) to create the image. This is necessary because we first want to run the initial image to create a checkpoint, the actual production image is created by committing the container of that run to an image.