@SleeplessByte 2025-03-13
Ahhh yes. @fapdash that makes sense.

@ErikSchierboom is there already a pattern for this in another track? IMO it makes sense what's being proposed here...