@ErikSchierboom 2025-03-13
> @ErikSchierboom is there already a pattern for this in another track? IMO it makes sense what's being proposed here...

Not really. I'll have to take some time to look at this, but maybe @iHiD can chip in to see if he wants to support those extra flags needing to be passed in.