@fapdash 2025-03-16
@ErikSchierboom I think I found a way to build the image without changing `exercism/github-actions/.github/workflows/docker-build-push-image.yml`. It's inspired by Micronauts CRaC Gradle plugin.

The new [deploy action](https://github.com/exercism/java-test-runner/blob/16612d31c7045766b9ef7864c6a2f91c8114b3d8/.github/workflows/deploy.yml) works as follows:
1. Build an image with an entrypoint that creates a CRaC checkpoint. This is now done via `docker/build-push-action`, so the image gets cached. Uses [Dockerfile.createCheckpoint](https://github.com/exercism/java-test-runner/blob/16612d31c7045766b9ef7864c6a2f91c8114b3d8/Dockerfile.createCheckpoint)
2. Run the image as a container to create the checkpoint. The checkpoint gets written into a directory on the host.
3. Run Exercisms `docker-build-push-image.yml` action. The Docker build copies the checkpoint generated in step 2 from the host file system. Uses a different [Dockerfile](https://github.com/exercism/java-test-runner/blob/16612d31c7045766b9ef7864c6a2f91c8114b3d8/Dockerfile) than step 1

I'm unsure if it'll be a problem that we [set up buildx](https://github.com/exercism/java-test-runner/blob/6eb9f473631675e9f84d89fc6a1845caeb236351/.github/workflows/deploy.yml#L28-L29) twice, once in this repos deploy action and then inside the `docker-build-push-image` action that we're calling.

Please don't do a detailed review yet, there are still some minor todos: cleanup, documentation, strict version pinning in the GHAs.. But I'm interested in feedback regarding the new build approach. :)