@SleeplessByte 2025-03-19
> Do you still need me on this?

Not right now I think. If https://github.com/exercism/java-test-runner/pull/147#issuecomment-2727148167 works, no flags are needed.