@fapdash 2025-04-23
Meh, just realized that the deploy job already broke last month, it just doesn't get reported inside of the PR.
So my idea in https://github.com/exercism/java-test-runner/pull/147#issuecomment-2727148167 doesn't work. :(

It's not possible to call a reusable workflow inside of `steps`, it has to be called by a job, which in turn can't have any steps.

There's the possibility to split the build into two jobs, but the jobs don't share a common workspace. We'd have to first upload the artifacts (jar+checkpoint) and download them in the job calling the reusable workflow: https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/storing-and-sharing-data-from-a-workflow#passing-data-between-jobs-in-a-workflow
So we'd still have to change the GitHub action upstream.
This could be an alternative to https://github.com/exercism/github-actions/pull/208, but I'm not sure if it offers any benefits? Wdyt?