@fapdash 2025-09-24
@IsaacG I'm still interested in getting this PR merged, but we'd need a decision on if it's worth the effort changing the GHA upstream. My time is very limited at least until the end of the year and I don't want to put additional work in that might be discarded. See my previous comment [here](https://github.com/exercism/java-test-runner/pull/147#issuecomment-2824954748) and Eriks comment [here](https://github.com/exercism/github-actions/pull/208#issuecomment-2726322989).